requirements.txt
# ─── API ────────────────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic==2.7.1

# ─── Data Science & ML ──────────────────────────────────────────────────────────
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
joblib==1.4.0

# ─── Database ───────────────────────────────────────────────────────────────────
mysql-connector-python==9.0.0
sqlalchemy==2.0.29

# ─── Utilitaires ────────────────────────────────────────────────────────────────
python-dotenv==1.0.1
httpx==0.27.0
watchdog==4.0.0
